support_backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
prometheus-fastapi-instrumentator
psycopg2-binary